requirements.txt
langchain
langchain-groq
langchain-community
langgraph
python-dotenv
ollama
langchain-ollama
langchain-core